+++ src/views/review/AddReview.vue
@@ -17,23 +17,11 @@
         ></v-rating>
       </v-col>
     </v-row>
-
     <v-row>
       <v-col class="d-none d-sm-flex" cols="2" md="3"></v-col>
       <v-col cols="12" sm="8" md="6" align-self="center">
         <v-form ref="form" lazy-validation class="mt-5 mb-10" v-model="valid">
-          <AlertDialog
-            :d_title="posts.d_title"
-            :d_value="posts.d_value"
-            :d_message="posts.d_message"
-            :d_cancel="posts.d_cancel"
-            :d_accept="posts.d_accept"
-            :d_color="posts.d_color"
-            :d_icon="posts.d_icon"
-            @dialog-accept="acceptButtonDialog(posts.id)"
-            @dialog-cancel="posts.d_value = false"
-          />
-
+          
           <v-textarea
             v-model="review.r_description"
             label="Descripción"
@@ -71,14 +59,12 @@
 
 <script>
 import { Review } from "/firebaseAPI/controllers/review.js";
-import AlertDialog from "@/components/Dialog.vue";
 import { getAuth } from "firebase/auth";
 
 export default {
-  name: "AddProduct",
-  components: {
-    AlertDialog,
-  },
+  name: "AddReview",
+  components: {},
+
   data() {
     return {
       rating: 1,
@@ -89,45 +75,9 @@ export default {
         maxstars: 5,
         disabled: false,
       },
-      posts: {
-        id: null,
-        d_value: false,
-        d_title: "",
-        d_message: "",
-        d_cancel: "",
-        d_accept: "",
-        d_color: "",
-        d_icon: "",
-      },
-      dialog: [
-        {
-          //Dialogo: Tamaño de Imagen
-          id: 0,
-          d_value: true,
-          d_title: "Image Size Exceeded",
-          d_message: "Select an image with a smaller size",
-          d_cancel: "Ok",
-          d_accept: "Ok",
-          d_color: "error",
-          d_icon: "fa fa-times",
-        },
-        {
-          //Dialogo: (Prueba) Modificar Producto
-          id: 1,
-          d_value: true,
-          d_title: "Edit Product",
-          d_message: "Are you sure to edit this product?",
-          d_cancel: "Ok",
-          d_accept: "Yes, update",
-          d_color: "primary",
-          d_icon: "fa fa-edit",
-        },
-      ],
+    
       valid: true,
-      rules: [
-        (value) =>
-          !value || value.size < 2000000 || "El archivo debe ser menor a 2MB",
-      ],
+      
       review: {
         id_user: 1,
         id_business: 1,
@@ -136,13 +86,12 @@ export default {
         r_datetime: "",
         r_status: false,
       },
-      items: ["Comida", "Stickers", "Dulces", "Bebidas", "Ropa", "Panaderia"],
+     
     };
   },
 
-  async created() {
-    // Validar que el usuario esté loggeado
-  },
+  async created() {},
+
   methods: {
     getStars(stars) {
       this.star = stars;
@@ -154,66 +103,18 @@ export default {
         this.review.id_business = 1;
         this.review.r_rate = this.rating;
         this.review.r_datetime = this.getDate();
-        review.addReview(this.review).then().catch();
+        await review.addReview(this.review).then().catch();
         this.$router.push("Review");
       } else {
-        // 0 -> Dialogo de imagen grande
-        // 1 -> (Prueba) Dialogo editar producto
-        this.showDialog(0);
+        //
       }
     },
     getDate() {
       const today = new Date();
-      const months = [
-        "Enero",
-        "Febrero",
-        "Marzo",
-        "Abril",
-        "Mayo",
-        "Junio",
-        "Julio",
-        "Agosto",
-        "Septiembre",
-        "Octubre",
-        "Noviembre",
-        "Diciembre",
-      ];
-
-      const daysOfWeek = [
-        "Domingo",
-        "Lunes",
-        "Martes",
-        "Miércoles",
-        "Jueves",
-        "Viernes",
-        "Sábado",
-      ];
-
-      return `${today.getDate()}/${
-        months[today.getMonth()]
-      }/${today.getFullYear()}`;
+      return `${today.getFullYear()}-${today.getMonth().toString()}-${today.getDate().toString()} ${
+        today.getHours().toString()}:${today.getMinutes().toString()}:${today.getSeconds().toString()}`;
     },
 
-    showDialog(id) {
-      this.posts.id = this.dialog[id].id;
-      this.posts.d_value = this.dialog[id].d_value;
-      this.posts.d_title = this.dialog[id].d_title;
-      this.posts.d_message = this.dialog[id].d_message;
-      this.posts.d_cancel = this.dialog[id].d_cancel;
-      this.posts.d_accept = this.dialog[id].d_accept;
-      this.posts.d_color = this.dialog[id].d_color;
-      this.posts.d_icon = this.dialog[id].d_icon;
-    },
-    acceptButtonDialog(id) {
-      switch (id) {
-        case 0:
-          this.$refs.image.reset();
-          this.posts.d_value = false;
-          break;
-        default:
-          break;
-      }
-    },
     goBack() {
       this.$router.back();
     },
